requirements_unspecified.txt
bcrypt
Cython
pycryptodome
PySide6
PySide6-Addons
PySide6-Essentials
python-dotenv
shiboken6
unqlite
python-dotenv
